requirements.txt
numpy
pandas
plotly
plotly-express
scikit-learn
kaleido
nbformat
openpyxl
pillow
matplotlib
